chore: Move "Update Track" logic over
